C_1:Terrain and Texture
This commit involves adding a terrain of 400x400 with max height of 100. Some mountains are added. Textures used to paint are: Rock, Snow, Light Grass and Grass Hill. Base Layer is Light Grass with some spots having Grass Hill. Mountains are painted with Rock with Snow covering the top. Some snow exists in the entire scene giving it a wintery feel